requirements.txt
pyttsx3==2.90
cyhunspell==2.0.2
Keras==2.4.3
SpeechRecognition==3.8.1
numpy==1.19.5
opencv_python==4.5.1.48
tensorflow==2.4.1
hunspell==0.5.5
Pillow==8.2.0
